Implement simple example using swash to render to png (#54)
This is basically working. But it could definitely do with some actual
review and cleanup before merging.

## Output


![output](https://github.com/linebender/parley/assets/1007307/f18b2850-0638-496b-93c6-3c2aa9c90f2b)

## Notes

- ~Emoji rendering is not working~
- Font fallback for arabic text does seem to be working (note: I pulled
the arabic text from cosmic text's sample files. Google translate
translate the included arabic text as "arabic", but I cannot otherwise
vouch for it's content).
- Subpixel rendering is not enabled (does this make sense when rendering
to image?)
- The rendered text doesn't look fantastic. Possibly my blending code
(used to apply the Swash alpha mask to an RGBA ImageBuffer from the
image buffer is incorrect / too naive?). Cosmic-text pulls in tiny-skia
in it's examples. I was hoping to avoid that here, but perhaps it's
necessary?
- ~Converting from a `parley::Font` to a `swash::FontRef` is very
awkward, specifically when the font file is a collection. The Parley
`Font` stores the index of the font within the collection, but Swash's
`FontRef` expect the byte offset of the font within the collection. I
could not find this exposed anywhere in
parley/fontique/skrifa/read-fonts so I'm currently doing pointer
arithmetic to get it.~
- ~I'm currently writing to a file called `output.png` in the users
current working directory. Possibly we could do better than that (some
kind of gitignored "data" directory in the parley repo?)~
- ~I've added an impl of `Brush` for `peniko::Color`, but possibly we
ought to just blanket impl for all types implementing the requisite
super traits?~
- ~I'm not using the `NormalizedCoords` as they always seemed to be
empty in my example (what are they for?)~