[parley] new selection logic and an editor example (#106)
Adds a new `Selection` type for dealing with cursor movement and
selection.

Also includes a `vello_editor` example that serves as a testbed for
checking the logic. This supports mouse selections, arrow keys (+ shift
to extend selections), backspace, delete and basic text input.

Marked as draft because the code is a mess but making this public so
people can play with it and report bugs.